requirements.txt
streamlit
pandas
numpy
langchain>=0.1.13
faiss-cpu
langchain-google-genai
google-generativeai
sentence-transformers
transformers
